+++ gik-api/src/routers/router.go
@@ -10,7 +10,7 @@ package routers
 // import "github.com/gin-gonic/gin"
 
 import (
-	"GIK_Web/src/routers/category"
+	"GIK_Web/src/routers/classification"
 	"GIK_Web/src/routers/status"
 
 	"github.com/gin-gonic/gin"
@@ -22,9 +22,9 @@ func InitRouter() *gin.Engine {
 
 	r.GET("/ping", status.Ping)
 
-	categoryApis := r.Group("/category")
+	categoryApis := r.Group("/classification")
 	{
-		categoryApis.GET("/list-category", category.ListCategory)
+		categoryApis.GET("/list-category", classification.ListCategory)
 	}
 
 	return r
